Add files via upload
Due to the increasing use of smart phones for reasons such as ease of transportation and ease of use in the user system, it has increased and stored personal data on these devices. Ransomware, which is a type of malicious program, demands a ransom payment for the return of information after infecting the relevant device, and these types of malware encrypt a user's personal data to prevent a person from accessing healthy data. In this research, we provide a model for an Android ransomware intrusion detection system that improves the previous methods of detecting Android ransomware.This was done through adequate data preprocessing using machine learning classifiers, decision tree, random forest and XGboost. Network traffic data were used for this classification. Of the three classifiers, XGboost produced better results.